feat(movement): Use physics for jumping
Rewrite jumping logic to use Unity physics engine
Fix AbsorbShield causing player to get boosted up due to new physics
